Moved 3 funcs to `sprites.py`: update_sprite, update_mask, generate_sprite.
