+++ scripts/cat/cats.py
@@ -46,9 +46,8 @@ from scripts.game_structure.screen_settings import screen
 from scripts.housekeeping.datadir import get_save_dir
 from scripts.utility import (
     clamp,
-    update_sprite,
-    update_mask,
 )
+from scripts.cat.sprites import update_sprite, update_mask
 from scripts.events_module.text_adjust import (
     event_text_adjust,
     leader_ceremony_text_adjust,
