+++ scripts/screens/SpriteInspectScreen.py
@@ -8,14 +8,13 @@ from scripts.cat.cats import Cat
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import UIImageButton, UISurfaceImageButton
 from scripts.utility import (
-    generate_sprite,
     get_text_box_theme,
 )
 from ..events_module.text_adjust import shorten_text_to_fit
 from ..ui.scale import ui_scale, ui_scale_dimensions, ui_scale_offset
 from .Screens import Screens
 from .enums import GameScreen
-from ..cat.sprites import sprites
+from ..cat.sprites import sprites, generate_sprite
 from .enums import GameScreen
 from ..clan_package.settings import get_clan_setting
 from ..game_structure.game.switches import switch_set_value, switch_get_value, Switch
